Make auto-config of RestTemplate metrics back off with no builder bean
See https://github.com/spring-projects/spring-boot/issues/15719
